sql: move time zone functions to a separate file
